First Agent/requirements.txt
claude-agent-sdk>=0.1.19
python-dotenv>=1.0.0
aiohttp>=3.9.0
aiosqlite>=0.19.0
